Renaming to lib/util/patterns/letters.js
