+++ lib/rules/capitalized-comments.js
@@ -8,7 +8,7 @@
 // Requirements
 //------------------------------------------------------------------------------
 
-const LETTER_PATTERN = require("../util/patterns/L");
+const LETTER_PATTERN = require("../util/patterns/letters");
 
 //------------------------------------------------------------------------------
 // Helpers
