+++ lib/util/patterns/letters.js
